+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Newer
 -----
 * Form Element Select: suppress placeholder (by setting option to false)
 * Form Element Checkbox: support option 'presets' (and 'presets:label')
+* Form Element Array: new option 'index_type' (keep, _keys or array)
 
 2017-09-15 (Version >= 2.6)
 ---------------------------